backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
python-multipart
requests
google-genai
pgvector
alembic
tenacity
pydantic-settings
python-dotenv
tiktoken
aiohttp
fpdf2
numpy
scikit-learn